Bugfix - ReadGenepopHWE update exact test for v4.7
In Genepop v4.7 the "Exact" test option no longer reports "High. sign." for overall pop and loci results, instead it reports " < 0.1555" or whatever the min pvalue possible is.

This update also includes a new flag, 'summaryAsNumeric', which defaults to TRUE. When set to TRUE, for Exact test output, the SummaryPValues matrix will be numeric. "High. sign." -> 0.0000 and " < 0.1555" -> 0.1555. If set to FALSE, then the SummaryPValues matrix will be character and preserve "High. sign." and " < 0.1555" values.